moved analyzer to its own module (#55)
